Move onboarding provider outside `onboarding/layout.tsx`
